requirements.txt
streamlit>=1.33.0
python-pptx>=1.0.0
lxml>=5.0.0
Pillow>=10.0.0
pywin32>=306
pydantic>=2.0.0
Jinja2>=3.1.0
st-copy-to-clipboard>=0.1.0
pyyaml>=6.0.0
pytest>=8.0.0
